back-end/requirements.txt
# ============================================
# REQUIREMENTS BACKEND - Flask API with SQLAlchemy
# ============================================
# Production-ready dependencies untuk Clean Architecture

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database & ORM
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5          # Database migrations
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication & Security
Flask-JWT-Extended==4.5.3     # JWT tokens untuk auth
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation & Serialization
marshmallow==3.20.1           # Object serialization/validation
flask-marshmallow==0.15.0

# Image Processing
Pillow==11.0.0
Werkzeug==3.0.1

# Dependency Injection
# Note: dependency-injector requires C++ compiler on Windows
# Consider installing without it and using manual DI instead
# dependency-injector==4.42.0

# Utilities
python-dateutil==2.8.2

# Development & Testing
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0                # Code formatter
flake8==7.1.1                 # Linter
